refactor the search functionality working process
Instead of have a local state of the images, add a displayedImages prop
that will be used to render and a storedImages prop to keep track of the data
to the redux store. When a user types to search for an image updatge
the displayedImages by comparing it to the storedImages.